+++ src/pyprediktormapclient/__init__.py
@@ -1,4 +1,7 @@
 import sys
+from .common import *
+from .model_index import *
+from .opc_ua import *
 
 if sys.version_info[:2] >= (3, 8):
     # TODO: Import directly (no need for conditional) when `python_requires = >= 3.8`
